requirements.txt
PyYAML
aws-cdk.aws-s3==1.99.0
aws-cdk.aws-lambda==1.99.0

flake8